Removing old form that was not needed
Instead of inheriting, I'm moving the code to the appropriate
namespace.
